PySide2.QtCore.QPoint — Type

QPoint

Type that defines a point in the plane using integer precision. It is the representation of PySide.QtCore.QPoint. A point is specified by its x and y coordinates.

Fields

xpos – Object x-position

ypos – Object y-position

PySide2.QtCore.QSize — Type

QSize

Type that defines the size of a two-dimensional object using integer point precision. It is the representation of PySide2.QtCore.QSize. QSize types can be summed, subtracted or tested for equality.

Fields

width – Object width

height – Object height

PySide2.QtCore.QLine — Type

QLine

Type that provides a two-dimensional vector using integer precision. It is the representation of PySide.QtCore.QLine. It describes a finite length line on a two-dimensional surface. The start and end points of the line are specified using integer point accuracy for coordinates. Use the QlineF constructor to retrieve a floating point copy.

Fields

x1 – initial point x-coordinate

y1 – initial point y-coordinate

x2 – final point x-coordinate

y2 – final point y-coordinate

PySide2.QtCore.QMargins — Type

QMargins

Type that defines the four margins of a rectangle. It is the representation of PySide.QtCore.QMargins. Thse margins describe the size of the borders surrounding a rectangle.

Fields

left – left margin

right – right margin

top – top margin

bottom – bottom margin

PySide2.QtCore.QRect — Type

QRect

Type that defines a rectangle in the plane using integer precision. It is the representation of PySide.QtCore.QRect. A rectangle is normally expressed as a top-left corner and a size. The size (width and height) of a QRect is always equivalent to the mathematical rectangle that forms the basis for its rendering.

A QRect can be constructed with a set of left, top, width and height integers, or from a QPoint and a QSize.

Fields

top – Object x-position

left – Object y-position

width – Rectangle width

height – Rectangle height

PySide2.QtWidgets.QWidget — Type

QWidget

Abstract type interface for the QWidget base class. QWidget is the atom of the user interface: it receives mouse, keyboard and other events from the window system, and paints a representation of itself on the screen. Every widget is rectangular, and they are sorted in a Z-order. A widget is clipped by its parent and by the widget in front of it.

Notes

For a compact Julia interface, each concrete subtype of QWidget has to inherit the Python class object in a field named obj.

PySide2.QtWidgets.QMainWindow — Type

QWindow <: QWidget

PySide2.QtWidgets.QLabel — Type

QLabel

The QLabel widget provides a text or image display. It is the julia interface to the PySide2.QtWidgets.QLabel class. No user interaction functionality is provided. The visual appearance of the label can be configured in various ways, and it can be used for specifying a focus mnemonic key for another widget.

PySide2.QtWebEngineWidgets.QWebEngineView — Type

QWebEngineView

This type is a julia interface for the QWebEngineView class, which provides a widget that is used to view and edit web documents. A web view is the main widget component of the Qt WebEngine web browsing module. It can be used in various applications to display web content live from the Internet.